+++ luxonis_ml/nn_archive/config_building_blocks/base_models/__init__.py
@@ -1,11 +1,6 @@
 from .head import (
     Head,
-    HeadClassification,
-    HeadObjectDetection,
-    HeadObjectDetectionSSD,
-    HeadSegmentation,
     HeadType,
-    HeadYOLO,
 )
 from .input import Input, PreprocessingBlock
 from .metadata import Metadata
@@ -14,11 +9,6 @@ from .output import Output
 __all__ = [
     "Head",
     "HeadType",
-    "HeadSegmentation",
-    "HeadClassification",
-    "HeadObjectDetection",
-    "HeadObjectDetectionSSD",
-    "HeadYOLO",
     "Input",
     "PreprocessingBlock",
     "Output",
