+++ luxonis_ml/nn_archive/config_building_blocks/base_models/head.py
@@ -1,22 +1,15 @@
 from abc import ABC
-from typing import Optional, Union
+from typing import List, Optional, Union
 
-from pydantic import BaseModel, Field, model_validator
+from pydantic import BaseModel, Field
 
-from ..enums import ObjectDetectionSubtypeYOLO
 from .head_metadata import (
     HeadClassificationMetadata,
     HeadMetadata,
     HeadObjectDetectionMetadata,
+    HeadObjectDetectionSSDMetadata,
     HeadSegmentationMetadata,
     HeadYOLOMetadata,
-)
-from .head_outputs import (
-    Outputs,
-    OutputsClassification,
-    OutputsSegmentation,
-    OutputsSSD,
-    OutputsYOLO,
 )
 
 
@@ -25,224 +18,30 @@ class Head(BaseModel, ABC):
 
     @type parser: str
     @ivar parser: Name of the parser responsible for processing the models output.
-    @type postprocessor_path: str | None
-    @ivar postprocessor_path: Path to the postprocessor.
-    @type outputs: C{Outputs}
-    @ivar outputs: A configuration specifying which output names from the `outputs` block of the archive are fed into the head.
-    @type metadata: C{HeadMetadata}
+    @type outputs: List[str] | None
+    @ivar outputs: Specify which outputs are fed into the parser. If None, all outputs
+        are fed.
+    @type metadata: C{HeadMetadata} | C{HeadObjectDetectionMetadata} |
+        C{HeadClassificationMetadata} | C{HeadObjectDetectionSSDMetadata} |
+        C{HeadSegmentationMetadata} | C{HeadYOLOMetadata}
     @ivar metadata: Metadata of the parser.
     """
 
     parser: str = Field(
         description="Name of the parser responsible for processing the models output."
     )
-    postprocessor_path: Optional[str] = Field(description="Path to the postprocessor.")
-    metadata: HeadMetadata = Field(description="Metadata of the parser.")
-    outputs: Outputs = Field(
-        description="A configuration specifying which output names from the `outputs` block of the archive are fed into the head."
+    metadata: Union[
+        HeadMetadata,
+        HeadObjectDetectionMetadata,
+        HeadClassificationMetadata,
+        HeadObjectDetectionSSDMetadata,
+        HeadSegmentationMetadata,
+        HeadYOLOMetadata,
+    ] = Field(description="Metadata of the parser.")
+    outputs: Optional[List[str]] = Field(
+        None,
+        description="Specify which outputs are fed into the parser. If None, all outputs are fed.",
     )
 
 
-class HeadObjectDetection(Head, ABC):
-    """Metadata for object detection head.
-
-    @type metadata: HeadObjectDetectionMetadata
-    @ivar metadata: Metadata of the parser.
-    """
-
-    metadata: HeadObjectDetectionMetadata = Field(description="Metadata of the parser.")
-
-
-class HeadClassification(Head, ABC):
-    """Metadata for classification head.
-
-    @type outputs: C{OutputsClassification}
-    @ivar outputs: A configuration specifying which output names from the `outputs` block of the archive are fed into the head.
-    @type metadata: C{HeadClassificationMetadata}
-    @ivar metadata: Metadata of the parser.
-    """
-
-    outputs: OutputsClassification = Field(
-        description="A configuration specifying which output names from the `outputs` block of the archive are fed into the head."
-    )
-    metadata: HeadClassificationMetadata = Field(description="Metadata of the parser.")
-
-
-class HeadObjectDetectionSSD(HeadObjectDetection, ABC):
-    """Metadata for SSD object detection head.
-
-    @type outputs: C{OutputsSSD}
-    @ivar outputs: A configuration specifying which output names from the `outputs` block of the archive are fed into the head.
-    @type metadata: C{HeadObjectDetectionMetadata}
-    @ivar metadata: Metadata of the parser.
-    """
-
-    outputs: OutputsSSD = Field(
-        description="A configuration specifying which output names from the `outputs` block of the archive are fed into the head."
-    )
-    metadata: HeadObjectDetectionMetadata = Field(description="Metadata of the parser.")
-
-
-class HeadSegmentation(Head, ABC):
-    """Metadata for segmentation head.
-
-    @type outputs: C{OutputsSegmentation}
-    @ivar outputs: A configuration specifying which output names from the `outputs` block of the archive are fed into the head.
-    @type metadata: C{HeadSegmentationMetadata}
-    @ivar metadata: Metadata of the parser.
-    """
-
-    outputs: OutputsSegmentation = Field(
-        description="A configuration specifying which output names from the `outputs` block of the archive are fed into the head."
-    )
-    metadata: HeadSegmentationMetadata = Field(description="Metadata of the parser.")
-
-
-class HeadYOLO(HeadObjectDetection, HeadSegmentation, ABC):
-    """Metadata for YOLO head.
-
-    @type outputs: C{OutputsYOLO}
-    @ivar outputs: A configuration specifying which output names from the `outputs` block of the archive are fed into the head.
-    @type metadata: C{HeadYOLOMetadata}
-    @ivar metadata: Metadata of the parser.
-    """
-
-    outputs: OutputsYOLO = Field(
-        description="A configuration specifying which output names from the `outputs` block of the archive are fed into the head."
-    )
-    metadata: HeadYOLOMetadata = Field(description="Metadata of the parser.")
-
-    @model_validator(mode="before")
-    def validate_task_specific_fields(
-        cls,
-        values,
-    ):
-        defined_params = {
-            k for k, v in (values["metadata"].model_dump()).items() if v is not None
-        }
-
-        common_fields = [
-            "subtype",
-            "iou_threshold",
-            "conf_threshold",
-            "max_det",
-            "classes",
-            "n_classes",
-            "anchors",
-        ]
-        defined_params = defined_params.difference(common_fields)
-
-        required_fields = {
-            "instance_segmentation": [
-                "n_prototypes",
-                "is_softmax",
-            ],
-            "keypoint_detection": ["n_keypoints"],
-        }
-
-        unsupported_fields = {
-            "object_detection": [
-                "n_prototypes",
-                "n_keypoints",
-                "is_softmax",
-            ],
-            "instance_segmentation": ["n_keypoints"],
-            "keypoint_detection": [
-                "n_prototypes",
-                "is_softmax",
-            ],
-        }
-
-        tasks = []
-        # Extract the task type
-        if all(
-            [
-                field in defined_params
-                for field in required_fields.get("instance_segmentation", [])
-            ]
-        ):
-            tasks.append("instance_segmentation")
-        if all(
-            [
-                field in defined_params
-                for field in required_fields.get("keypoint_detection", [])
-            ]
-        ):
-            tasks.append("keypoint_detection")
-        if all(
-            [
-                field not in defined_params
-                for field in unsupported_fields.get("object_detection", [])
-            ]
-        ):
-            tasks.append("object_detection")
-
-        if len(tasks) == 0:
-            raise ValueError(
-                "Invalid combination of parameters. No specific task can be inferred."
-            )
-
-        for param in defined_params:
-            if not any(param in required_fields[task] for task in tasks):
-                raise ValueError(
-                    f"Invalid combination of parameters. Field {param} is not supported for the tasks {tasks}."
-                )
-
-        # Validate Outputs
-        outputs = values.get("outputs", {})
-        defined_params = {
-            k
-            for k, v in (
-                outputs.model_dump() if isinstance(outputs, BaseModel) else outputs
-            ).items()
-            if v is not None
-        }
-
-        supported_output_params = {
-            "instance_segmentation": ["yolo_outputs", "mask_outputs", "protos"],
-            "keypoint_detection": ["yolo_outputs", "keypoints"],
-            "object_detection": ["yolo_outputs"],
-        }
-
-        # Check if all required output fields are present
-        if not all(
-            [
-                field in defined_params
-                for task in tasks
-                for field in supported_output_params[task]
-            ]
-        ):
-            raise ValueError(f"Invalid output fields for tasks {tasks}")
-
-        # Check if all defined fields are supported
-        for param in defined_params:
-            if param == "angles" and "object_detection" in tasks:
-                continue
-            if not any(param in supported_output_params[task] for task in tasks):
-                raise ValueError(
-                    f"Invalid combination of output parameters. Field {param} is not supported for the tasks {tasks}."
-                )
-
-        return values
-
-    @model_validator(mode="before")
-    def validate_anchors(cls, values):
-        if (
-            "anchors" in values
-            and values["anchors"] is not None
-            and (
-                values["subtype"] == ObjectDetectionSubtypeYOLO.YOLOv6
-                or values["subtype"] == ObjectDetectionSubtypeYOLO.YOLOv8
-            )
-        ):
-            raise ValueError("YOLOv6 and YOLOv8 do not support anchors.")
-        return values
-
-
-HeadType = Union[
-    HeadClassification,
-    HeadObjectDetection,
-    HeadObjectDetectionSSD,
-    HeadSegmentation,
-    HeadYOLO,
-]
+HeadType = Head
